Try to install the latest version if there are multiple versions for the same package
Prefer packages with no digits e.g. "llvm", since such packages seems to be the latest version.
If there is no package without digits, extract digits separated by dash (-) and dot (.),
and build an int array to represent the version, and pick the higest one.
For instance, "llvm-6.0" will be parsed as [6,0] and "llvm-10" as [10].